refactor(npm-scripts): move tweakWebpackConfig.js back to utils
Since we don't have so much webpack stuff, we can keep everything under
utils as before.
