+++ projects/npm-tools/packages/npm-scripts/src/utils/tweakWebpackConfig.js
